requirements.txt
anyio==4.4.0
fastapi==0.111.0
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic-core==2.20.1
sniffio==1.3.1
starlette==0.37.2
typing-extensions==4.12.2
colorama==0.4.6
python-dotenv==1.0.1
uvicorn==0.30.1
watchfiles==0.22.0
asyncpg==0.29.0
greenlet==3.0.3
SQLAlchemy==2.0.31
bcrypt==3.2.2
cryptography==42.0.8
passlib==1.7.4
python-jose==3.3.0
ecdsa==0.19.0
pydantic-settings==2.3.4